refactor driver code and models
