+++ driver/driver.go
@@ -1,0 +1,37 @@
+package driver
+
+import (
+	"database/sql"
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/lib/pq"
+)
+
+var db *sql.DB
+
+func ConnectDB() *sql.DB {
+	// DB setup
+	pgUrl, err := pq.ParseURL(os.Getenv("ELEPHANTSQL_URL"))
+
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	db, err = sql.Open("postgres", pgUrl)
+
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	fmt.Println(pgUrl)
+
+	err = db.Ping()
+
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	return db
+}
