+++ main.go
@@ -9,54 +9,23 @@ import (
 	"os"
 	"strings"
 
-	"github.com/subosito/gotenv"
-
-	// jwt "github.com/dgrijalva/jwt-go"
+	"./driver"
+	"./models"
 	"github.com/davecgh/go-spew/spew"
 	"github.com/dgrijalva/jwt-go"
 	"github.com/gorilla/mux"
-	"github.com/lib/pq"
+	"github.com/subosito/gotenv"
 	"golang.org/x/crypto/bcrypt"
 )
+
+var db *sql.DB
 
 func init() {
 	gotenv.Load()
 }
 
-type User struct {
-	ID       int32  `json:"id"`
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
-
-type JWT struct {
-	Token string `json:"token"`
-}
-
-type Error struct {
-	Message string `json:"message"`
-}
-
-var db *sql.DB
-
 func main() {
-
-	// DB setup
-	pgUrl, err := pq.ParseURL(os.Getenv("ELEPHANTSQL_URL"))
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	db, err = sql.Open("postgres", pgUrl)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println(pgUrl)
-
-	err = db.Ping()
+	db = driver.ConnectDB()
 
 	router := mux.NewRouter()
 	router.HandleFunc("/signup", signup).Methods("POST")
@@ -68,7 +37,7 @@ func main() {
 	log.Fatal(http.ListenAndServe(":8000", router))
 }
 
-func respondWithError(w http.ResponseWriter, status int, error Error) {
+func respondWithError(w http.ResponseWriter, status int, error models.Error) {
 	w.WriteHeader(status)
 	json.NewEncoder(w).Encode(error)
 }
@@ -77,7 +46,7 @@ func responseJSON(w http.ResponseWriter, data interface{}) {
 	json.NewEncoder(w).Encode(data)
 }
 
-func GenerateToken(user User) (string, error) {
+func GenerateToken(user models.User) (string, error) {
 	secret := os.Getenv("SECRET")
 
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
@@ -93,8 +62,8 @@ func GenerateToken(user User) (string, error) {
 }
 
 func signup(w http.ResponseWriter, r *http.Request) {
-	var user User
-	var error Error
+	var user models.User
+	var error models.Error
 
 	json.NewDecoder(r.Body).Decode(&user)
 
@@ -136,9 +105,9 @@ func signup(w http.ResponseWriter, r *http.Request) {
 
 func login(w http.ResponseWriter, r *http.Request) {
 
-	var user User
-	var jwt JWT
-	var error Error
+	var user models.User
+	var jwt models.JWT
+	var error models.Error
 
 	json.NewDecoder(r.Body).Decode(&user)
 
@@ -200,7 +169,7 @@ func protectedEndpoint(w http.ResponseWriter, r *http.Request) {
 
 func TokenVerifyMiddleWare(next http.HandlerFunc) http.HandlerFunc {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		var errorObject Error
+		var errorObject models.Error
 		authHeader := r.Header.Get("Authorization")
 		bearerToken := strings.Split(authHeader, " ")
 		// fmt.Println(bearerToken)
