+++ models/error.go
@@ -1,0 +1,5 @@
+package models
+
+type Error struct {
+	Message string `json:"message"`
+}
