+++ models/jwt.go
@@ -1,0 +1,5 @@
+package models
+
+type JWT struct {
+	Token string `json:"token"`
+}
